Moves code to src
