+++ package.json
@@ -2,10 +2,10 @@
   "name": "password-validator",
   "version": "2.1.2",
   "description": "Validates password according to flexible and intuitive specifications",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
-    "docs": "node_modules/.bin/jsdoc ./*.js -t ./node_modules/docdash -R README.md -d ./docs -r",
-    "lint": "eslint ./*.js",
+    "docs": "node_modules/.bin/jsdoc ./src/*.js -t ./node_modules/docdash -R README.md -d ./docs -r",
+    "lint": "eslint ./src",
     "test:unit": "mocha tests/*.js",
     "test": "npm run lint && npm run test:unit",
     "start": "node index.js"
