+++ src/lib.js
